requirements.txt
scikit-learn
scikit-image
librosa
numpy
pillow
hashlib
transformers
opencv-python
pandas
lpips
torch
matplotlib